+++ dev/oneoffs/enrich_dataset_with_hard_negatives.py
@@ -62,21 +62,24 @@ def main():
     cand_pred_sitesum_df = cand_pred_sitesum_df[is_contained]
 
     # Remove any prediction that touches any truth annotation
-    idx1_to_idxs2 = util_gis.geopandas_pairwise_overlaps(true_sitesum_df, cand_pred_sitesum_df, predicate='intersects')
-    idxs2 = sorted(ub.unique(ub.flatten(idx1_to_idxs2.values())))
-    is_intersecting = np.array(ub.boolmask(idxs2, len(cand_pred_sitesum_df)))
-    cand_pred_sitesum_df = cand_pred_sitesum_df[~is_intersecting]
+    # idx1_to_idxs2 = util_gis.geopandas_pairwise_overlaps(true_sitesum_df, cand_pred_sitesum_df, predicate='intersects')
+    # idxs2 = sorted(ub.unique(ub.flatten(idx1_to_idxs2.values())))
+    # is_intersecting = np.array(ub.boolmask(idxs2, len(cand_pred_sitesum_df)))
+    # cand_pred_sitesum_df = cand_pred_sitesum_df[~is_intersecting]
 
     region_id_to_true_sitesum = dict(list(true_sitesum_df.groupby('region_id')))
     region_id_to_region = dict(list(true_region_df.groupby('region_id')))
 
     import kwplot
     import kwimage
-    import geopandas as gpd
+    import geopandas as gpd  # NOQA
     kwplot.autompl()
 
     # Visualize candidates
-    for region_id, pred_subdf in cand_pred_sitesum_df.groupby('region_id'):
+    region_id_to_pred_subdf = dict(list(cand_pred_sitesum_df.groupby('region_id')))
+    for region_id, pred_subdf in region_id_to_pred_subdf.items():
+
+        pred_subdf = region_id_to_pred_subdf[region_id]
 
         fig = kwplot.figure(fnum=region_id, doclf=True)
         ax = fig.gca()
@@ -85,16 +88,18 @@ def main():
         region_rows = region_id_to_region[region_id]
 
         # hack
-        pred_union_geom = pred_subdf['geometry'].geometry.unary_union
-        pred_union_df = gpd.GeoDataFrame({'geometry': [pred_union_geom]})
+        # pred_union_geom = pred_subdf['geometry'].geometry.unary_union
+        # pred_union_df = gpd.GeoDataFrame({'geometry': [pred_union_geom]})
+        # pred_final_df = pred_union_df
+        pred_final_df = pred_subdf
 
         true_color = kwimage.Color.coerce('kitware_green').as01()
         fp_color = kwimage.Color.coerce('kitware_red').as01()
         bound_color = kwimage.Color.coerce('kitware_blue').as01()
 
         region_rows.plot(ax=ax, facecolor='none', edgecolor=bound_color, alpha=0.9)
+        pred_final_df.plot(ax=ax, facecolor='none', edgecolor=fp_color, alpha=0.9)
         true_subdf.plot(ax=ax, facecolor='none', edgecolor=true_color, alpha=0.9)
-        pred_union_df.plot(ax=ax, facecolor='none', edgecolor=fp_color, alpha=0.9)
 
         ax.set_title(f'region {region_id} - hard negatives')
         fpath = f'hardneg_{region_id}.png'